네비게이션바 리렌더링 최적화
