+++ web-app/components/NavigationBar/style.ts
@@ -1,6 +1,7 @@
 import styled from '@emotion/styled';
 
-export const NavigationBarWrapper = styled.div`
+export const NavigationBarWrapper = styled.nav`
+  max-width: 500px;
   position: fixed;
   bottom: 0;
   width: 100%;
@@ -12,13 +13,17 @@ export const NavigationBarWrapper = styled.div`
   box-shadow: 0 -1px 2px 0 rgba(0, 0, 0, 0.1);
 `;
 
-export const ButtonWrapper = styled.div`
+type ButtonWrapperProps = {
+  active: boolean;
+};
+
+export const ButtonWrapper = styled.div<ButtonWrapperProps>`
   display: flex;
   flex-basis: 0;
   flex-grow: 1;
   justify-content: center;
   align-items: center;
-  background-color: #ffffff;
+  background-color: ${({ active }) => (active ? '#c7c7f9' : '#ffffff')};
   position: relative;
 
   > a {
@@ -30,13 +35,13 @@ export const ButtonWrapper = styled.div`
     padding: 14px 0;
 
     :active::before {
+      content: '';
       position: absolute;
       left: 0;
       top: 0;
       width: 100%;
       height: 100%;
       background-color: rgba(0, 0, 0, 0.2);
-      content: '';
     }
   }
 `;
